Bugfix: hopefully fixing delete of to much entries

+++ upload/inc/tasks/reservations.php
@@ -18,10 +18,11 @@ function task_reservations($task)
   $get_type = $db->simple_select("reservationstype", "*");
   //für jeden typen nacheinander schauen
   while ($entry = $db->fetch_array($get_type)) {
+    if ($entry['member_lock'] != 0) {
     //jetzt die abgelaufenen löschen
-    $db->delete_query("reservationsentry", "type = '{$entry['type']}' and DATE_ADD(enddate, INTERVAL {$entry['member_lock']} DAY) < CURDATE()");
-    $db->delete_query("reservationsmodread", "entry_id NOT IN (SELECT entry_id FROM ".TABLE_PREFIX."reservationsentry");
+      $db->delete_query("reservationsentry", "type = '{$entry['type']}' and DATE_ADD(enddate, INTERVAL {$entry['member_lock']} DAY) < CURDATE()");
+      $db->delete_query("reservationsmodread", "entry_id NOT IN (SELECT entry_id FROM " . TABLE_PREFIX . "reservationsentry");
+    }
   }
-  add_task_log($task, "Reservierungen bereinigt");
-
+      add_task_log($task, "Reservierungen bereinigt");
 }
